small fixes in the EF manager classes

+++ ParkingWebApi_REST/Managers/EFSensorsManager.cs
@@ -18,6 +18,10 @@ namespace WebApi_REST.Managers
 
         public Sensor Add(Sensor sensor)
         {
+            if (_context.Sensors.Find(sensor.SensorId) != null)
+            {
+                return null;
+            }
             var newSensor = _context.Sensors.Add(sensor);
             _context.SaveChanges();
             return newSensor.Entity;
@@ -26,13 +30,19 @@ namespace WebApi_REST.Managers
         public void Delete(int id)
         {
             var sensorToDelete = _context.Sensors.Find(id);
-            _context.Sensors.Remove(sensorToDelete);
-            _context.SaveChanges();
-
+            if (sensorToDelete != null)
+            {
+                _context.Sensors.Remove(sensorToDelete);
+                _context.SaveChanges();
+            }
         }
 
         public Sensor Edit(int id, Sensor newSensor)
         {
+            if (_context.Sensors.Find(newSensor.SensorId) == null)
+            {
+                return null;
+            }
             var updatedSensor = _context.Sensors.Update(newSensor);
             _context.SaveChanges();
             return updatedSensor.Entity;
